Move Auth part to actions and create common webhook utility
