+++ play/src/index.ts
@@ -4,3 +4,4 @@ export { IAccount, getAccount, createSession, commitSession, fusebitRequest, Req
 export { playWrightConfig } from './playwright.config';
 export * as Utilities from './utilities';
 export { validate as validateTotp, generate as generateTotpToken } from './totpUtility';
+export { waitForWebhook } from './webhooks';
