+++ play/src/webhooks.ts
@@ -1,0 +1,36 @@
+import { expect } from '@playwright/test';
+import { Constants, fusebitRequest, IAccount, RequestMethod } from './index';
+
+// Common utility for webhook tests
+export async function waitForWebhook(account: IAccount, filter: (entry: any) => boolean): Promise<void> {
+  // Wait for the webhook event to fire.
+  let cnt: number;
+  for (cnt = 10; cnt > 0; cnt--) {
+    // Get the contents of the webhook storage and validate if it's what the test is looking for.
+    const response = await fusebitRequest(
+      account,
+      RequestMethod.get,
+      `/storage/integration/${Constants.INTEGRATION_ID}/test/${Constants.CONNECTOR_ID}/webhook/*`,
+      {},
+      { version: 1 }
+    );
+    expect(response).toBeHttp({ statusCode: 200 });
+
+    if (response.body.total > 0) {
+      const entries = await Promise.all(
+        response.body.items.map((item: { storageId: string }) =>
+          fusebitRequest(account, RequestMethod.get, `/storage/${item.storageId}`, {}, { version: 1 })
+        )
+      );
+      // Check to see if any of the entries match
+      if (entries.some(filter)) {
+        // Good enough for now - mark the test a success and move on.
+        break;
+      }
+    }
+
+    await new Promise((resolve) => setTimeout(resolve, 1000));
+    cnt -= 1;
+  }
+  expect(cnt).toBeGreaterThan(0);
+}
